requirements.txt
numpy
Pillow
pyrtlsdr
luma.core
luma.lcd
flask
RPi.GPIO
spidev
